refactor(dashpaint): widen both glyph ids so neither type pads
`GlyphQuad` and `AtlasGlyph` each led with a `u16` glyph id in a struct whose
other members are 4 bytes, so rustc inserted two bytes after it. That is
FFI-safe — a C compiler inserts the same — but not FFI-explicit, and story
#578's rules for anything crossing a language seam call for padding to be
declared. `crates/dashscene-unity` asserted the hole rather than fixing it,
with a comment saying #578 would.

Both ids are `u32` now, which removes the hole instead of naming it. The
alternative, an explicit `_pad` member, would have been a public field
participating in the derived `PartialEq` — two otherwise-equal quads could
differ in it, which is the equality hazard #699 has just finished removing
from the empty ranges. That trade binds every future boundary-B member
narrower than the ones around it, so it is recorded in
`docs/decisions/sub-word-members-widen-rather-than-pad.md` rather than left
here.

Nothing moves. `GlyphQuad` stays 12 bytes and `AtlasGlyph` 36, and both float
rectangles keep the offsets they had, so a consumer's declaration of them is
unchanged and every golden is byte-identical. OpenType ids are 16-bit,
`dashscene-typeset` still carries them as `u16`, and the five producer sites
widen with `u32::from`.

What the `u16` gave for free was that an out-of-domain id was unrepresentable.
Widening gives that up, and an id above `u16::MAX` would place a row
`Atlas::glyph` can still be asked for, so a quad naming it would paint nothing
with no diagnostic — the silent drop P4 forbids. `Atlas::new` refuses it by
name now, beside the sorted-and-unique assertion it already carried.

`AtlasGlyph` was not named in the story. It has the identical defect —
`{u16, [f32; 4], [f32; 4]}` pads exactly as `{u16, f32, f32}` does — so
fixing one and leaving the other would have left the same hole with a record
saying it was closed.

The replacement test asserts the id member's own size, because that is what
has teeth. Measured, not assumed: an offset-only version of it passed with the
`u16` put back on both types, since both structs were always the size they are
now and the member after the id always sat at offset 4. Narrow either id
together with whatever must change for the workspace to compile and the
matching assertion fails; narrow one on its own and the compiler refuses it
first.

Refs #578. Refs #569.

+++ corpus/showcase/src/resources.rs
@@ -117,7 +117,7 @@ fn load_atlas(dir: &str) -> Atlas {
         .iter()
         .filter_map(|glyph| {
             Some(AtlasGlyph {
-                glyph_id: glyph.glyph_id,
+                glyph_id: u32::from(glyph.glyph_id),
                 plane_em: glyph.plane_em?,
                 atlas_px: glyph.atlas_px?,
             })
